ai-features/image-emotion/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.1.0
python-dotenv==1.0.0
pydantic==2.5.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
requests==2.31.0
asyncio==3.4.3
aiohttp==3.9.1
pandas==2.1.4
scikit-learn==1.3.2
matplotlib==3.8.2
PyYAML==6.0.1
# Added for ML training
joblib==1.3.2
python-multipart==0.0.6
